Refactoring and fixing formatting
